feat!: BREAKING Make WebVitals a Plugin (#45)
This removes all the optional/nullable stuff we have to add to parts of
Portals code. If a user wants or needs WebVitals output for their
portals, they can add it as an instance plugin on the portal.

This also makes initialContext on Android a JavascriptInterface
call instead of setting window.portalInitialContext on the window. I
personally noted some scenarios where initialContext was not always
there. I believe this due to the nature of how we had to
evaluateJavascript on the webview every time a url was loaded. The
JavascriptInterface method makes this always available.
---------

Co-authored-by: Carl Poole <carl@ionic.io>